refactor(webrtc): extract shared receive jitter buffer and fan-out into stream_recv
